// constants/notification.enum.ts
export enum NotificationTypeEnum {
  TASK_DUE_SOON = 'task/due-soon',
  TASK_ASSIGNED = 'task/assigned',
  TASK_OVERDUE = 'task/overdue',
  TASK_COMPLETED = 'task/completed',
  TASK_UPDATED = 'task/updated',
  TASK_DELETED = 'task/deleted',
  TASK_CLEARED = 'task/cleared',

  PROJECT_DELETED = 'project/deleted',
  PROJECT_UPDATED = 'project/updated',
}